scripts/requirements.txt
requests>=2.31.0
Pillow>=10.0.0
gTTS==2.5.1
google-auth-oauthlib>=1.2.0
google-auth>=2.29.0
